add extension for fractional EXP multipliers on BB

+++ src/Lobby.hh
@@ -120,7 +120,7 @@ struct Lobby : public std::enable_shared_from_this<Lobby> {
   Episode episode;
   GameMode mode;
   uint8_t difficulty; // 0-3
-  uint16_t base_exp_multiplier;
+  float base_exp_multiplier;
   float exp_share_multiplier;
   float challenge_exp_multiplier;
   std::string password;
